refactor: :truck: move components to related pages and sections
