+++ pages/(home)/background-avatars.tsx
